requirements.txt
requests
python-dotenv
polars
google-cloud-bigquery
pyarrow
aiohttp